pyproject.toml
[project]
name = "kuntur-detector-server"
version = "0.1.0"
description = "Servicio de detección de amenazas para negocios ecuatorianos"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "pydantic>=2.4.2",
    "pytest>=7.4.3",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.10.0",
    "isort>=5.12.0",
    "pytest-cov>=4.1.0"
]
